[applications] E: Add accless-maa Baseline
