--- applications/functions/escrow-xput/include/utils.h
@@ -1,5 +1,0 @@
-#pragma once
-
-#include <string>
-
-std::string base64decode(const std::string &data);
