--- applications/functions/escrow-xput/src/utils.cpp
@@ -1,14 +1,0 @@
-#include "utils.h"
-
-#include <boost/algorithm/string.hpp>
-#include <boost/archive/iterators/binary_from_base64.hpp>
-#include <boost/archive/iterators/transform_width.hpp>
-
-std::string base64decode(const std::string &data) {
-    using namespace boost::archive::iterators;
-    using It =
-        transform_width<binary_from_base64<std::string::const_iterator>, 8, 6>;
-    return boost::algorithm::trim_right_copy_if(
-        std::string(It(std::begin(data)), It(std::end(data))),
-        [](char c) { return c == '\0'; });
-}
